Added stub definitions of the following two properties to all system objects:
    * _keyword_argument_names
    * _mandatory_argument_names

(The properties currently return empty tuples.)

A couple of generalized ways of doing this with introspection appear to be available.
But the properties are implemented as stubs here because none of the generalized routines
is yet testing correctly.

Class authors should look at these attributes as placeholders that will implement at some point in future.