Removed ImmutableAbjadObject. Use AbjadObject instead.
Background:

There have been two orthogonal ideas about 'immutability' implemented in the abctools classes.
The first such concept is the implementation of __slots__(). The second such concept
relates to the blocking of __setitem__() (and __delitem__()). Each of these is discussed below.

Python objects usually do not implement slots. But Abjad components all implement slots.
(As do many other Abjad classes.) The reason for this is that complete Abjad scores
instantiate thousands (or even tens of thousands) of Abjad score components. This is
exactly the situation in which slots become desirable. This means that Abjad components
will continue to implement slots as they do now.

Blocking __setitem__, __delitem__ turns out to be necessary for things like ImmutableDictionary.
Why? Because slots are enough to block *attribute* setting but do nothing to block *item* setting.

So when we reflect on these two principles we realize that slots will be widely implemented
in the Abjad codebase while blocking __setitem__, __delitem__ will be necessary for only
a small number of classes.

                                         * * *

One last important detail:

The classes in abctools ALL IMPLEMENT SLOTS.

The reason for this is because Python classes DO NOT inherit slot-like behavior.
Suppose class Foo implements slots. And suppose Bar subclasses Foo. Then the following
will be true:

    * Bar will block attribute assignment if (and only if) Bar also implements slots.
    * Bar will NOT block attriubte assignment if Bar does not implement slots.

That is to say that the fact that parent class Foo implemenets slots has absolutely
no bearing on whether child class Bar implements slots.

So this should help make clear why all abctools classes implement slots:

All Abjad classes should inherit from a class implemented in abctools. Because of this
it's important that all abctools classes implement (potentially trivial) slots.
This affords subclasses THE OPTION to either implement slots or not.

(If, on the other hand, abctools classes did NOT implement slots, then subclasses would
be unable to implement slots!)

                                  * * *

Finally, it seems like that the comparison functionality implemented in these abctools classes ...

    * UnaryComparator
    * StrictComparator
    * FlexEqualityComparator

... might be better implemented as decorators.

Research remains for later to implement and test this approach.