Deprecating ImmutableAbjadObject.
Use AbjadObject instead.

The ImmutableAbjadObject was introduced because of a confusion as to what immutability should mean.
Should immutability mean slots? Or should immutability mean the blocking of setattr?

It turns out that slots will be enough for most Abjad objects.

(Note: the primary purpose of slots is not to grant immutability but instead to reduce
the memory overhead of objects that may be instantiated hundreds or thousands of times.)