All schemetools classes now implement __slots__().
This was not previously the case.

What was missing here was the realization that *all classes in the inheritance hierarchy*
must explicitly include a "__slots__ = (... )" assignment in order for the child class
to implement slots.

What was previously happening was that the (base) schemetools.Scheme class implemented
__slots__ but that none of the child classes did.

The child classes now all explicitly assign slots.