Codebase no longer relies on bool(Container()) to return true.
This is a good thing.

The definition of AbjadObject.__nonzero__() will be removed at some point in future.

There is a concern, however:

It is possible that Abjad users have been relying on bool(Container()) to return true.
If this is the case, then removing the definition of AbjadObject.__nonzero__()
will break existing end-user scores.

This will have to be documented clearly in the Abjad 2.8 release notes.

Something like ...

    Note: in previous versions of Abjad the statement "if container: ..." returned true.
    This was ALWAYS the case, even when 'container' was empty.

    As of Abjad 2.8 the statement "if container: ..." returns false when
    container is empty and returns true when 'container' is nonempty.

... or the equivalent.