Updated classfile section headers.
The order is now:

    ### CLASS ATTRIBUTES ###

    ### INITIALIZER ###

    ### SPECIAL METHODS ###

    ### PRIVATE PROPERTIES ###
    
    ### PRIVATE METHODS ###

    ### PUBLIC PROPERTIES ###

    ### PUBLIC METHODS ###

Abjad classes use few class attributes. But the presence of class attributes is likely to grow.
Where class attributes are included, put them in a section at the head of the class definition.

(All section headers are optional. Classes include only those required by definition.)

The initializer isn't identified with a section header in most of the codebase right now.
Include it in newly written files (especially if class attributes preceed the initializer).
Later a script will be used to add the initializer section header globally.

Finally, it will eventually be nice to identify properties as read-only versus read / write
everywhere in the API. This will give:

    ### READ-ONLY PRIVATE PROPERTIES ###

    ### READ / WRITE PRIVATE PROPERTIES ###

    ### READ-ONLY PUBLIC PROPERTIES ###

    ### READ / WRITE PUBLIC PROPERTIES ###

We won't both with this distinction until the Grand Doc Overhaul (when we'll find a
way to cause Sphinx to surface this distinction in the published API).