Added new timetokentools package.
The package implements nine concrete classes:

    timetokentools.NoteFilledTimeTokenMaker
    timetokentools.OutputBurnishedSignalFilledTimeTokenMaker
    timetokentools.OutputIncisedNoteFilledTimeTokenMaker
    timetokentools.OutputIncisedRestFilledTimeTokenMaker
    timetokentools.RestFilledTimeTokenMaker
    timetokentools.SignalFilledTimeTokenMaker
    timetokentools.TokenBurnishedSignalFilledTimeTokenMaker
    timetokentools.TokenIncisedNoteFilledTimeTokenMaker
    timetokentools.TokenIncisedRestFilledTimeTokenMaker


A time-token is taken to be either a list of notes and rests or else a tuplet containing notes and rests.

So the main idea behind the package is to implement a related family of rhyhthm-making classes.

Usage of each class is the same:
    
    1. configurate a time-token maker at initialization;
       the resulting object acts a type of fully-loaded rhythm-maker;
    2. then call the time-token maker on as many arbitrary sequences of duration tokens as desired.

Examples are provided at each entry in the API.

In addition, many more examples are provided in the test files.

Only one example is included here:

    Configure the time-token maker at initialization::

        abjad> pattern, denominator, prolation_addenda = [-1, 4, -2, 3], 16, [3, 4]
        abjad> maker = timetokentools.SignalFilledTimeTokenMaker(pattern, denominator, prolation_addenda)

    Then call the time-token maker on arbitrary duration tokens::

        abjad> duration_tokens = [(2, 8), (5, 8)]
        abjad> music = maker(duration_tokens)

    The resulting Abjad objects can be included in any score and the time-token
    make can be called indefinitely on other arbitrary sequences of duration tokens::

        abjad> music = sequencetools.flatten_sequence(music)
        abjad> staff = Staff(measuretools.make_measures_with_full_measure_spacer_skips(duration_tokens))
        abjad> measuretools.replace_contents_of_measures_in_expr(staff, music)

        abjad> f(staff)
        \new Staff {
            {
                \time 2/8
                \times 4/7 {
                    r16
                    c'4
                    r8
                }
            }
            {
                \time 5/8
                \fraction \times 5/7 {
                    c'8.
                    r16
                    c'4
                    r8
                    c'8.
                    r16
                }
            }
        }

Intuitively, what's going on here is that the pattern [-1, 4, -2, 3] combines with the denominator 16
to create a repeating signal equal to "r16 c'4 r8 c8." which is then tupletized by the prolation
addenda given as [3, 4]. Those prolation addenda are interpreted only at call-time.

What happens at call-time is that the duration tokens [(2, 8), (5, 8)] are passed to the time-token maker.
The duration tokens are pairs. This is important because the denominators of duration tokens are taken
to be important. (This means that (2, 8) and (1, 4) are completely different duration tokens.)
The denominators of duration tokens are used in interpretating prolation addenda.
In our example we have prolation addenda [3, 4]. The 3 of [3, 4] is combined with denominator 16
to add 3/16 of time to duration token (2, 8); this explains the 7:4 tuplet in the output.
Then the 4 in [3, 4] is combined with denominator 16 to add 4/16 of time to duration (5, 8);
this expalins the 7:5 tuplet in the output.

The contents of the tuplets will be seen to be the input signal "r16 c'4 r8 c8" repeated cyclically
as required and with the durations of notes and rests broken across duration token boundaries.