Rename:
    scr/devel/rename-public-helper ==>
    scr/devel/rename-public-function

The concept of a 'helper' is no longer present in the system.

(This was the ancient name for any function made publicly available through the API.)

This change does not need to be noted in notes for the next release of the system
because the change applies to developer tools only.

Also cleaned up the script and fixed a bug that was keeping it from writing
the Sphinx versionchanged tag.