Streamlined contexttools.get_context_marks_attached_to_any_improper_parent_of_component().
Only a few lines changed.

Previously the function was giving intermittently failing results for some totally
inexplicable reason.

But the function now tests as stable.

Note too that the function previously returned a set but now returns a tuple.

(Too small a change to mention in the release notes, however.)