Rename:
    tietools.get_tie_chains_in_expr() ==>
    tietools.get_nontrivial_tie_chains_masked_by_components()

The function does fancier work that merely getting tie chains.

Specifically, the function implements logic to return *incomplete* tie chains.

Why?

Because in certain cases we may want to fuse just some of the notes in a tie chain together.

So the rename introduces the idea of a tie chain that is 'masked' in the sense of the Adobe
Illustrator or Photoshop masks. That is, the things returned by this function are tie chains
that may be 'masked' so that only part of the tie chain components appear.

Usage example will be added to the API in the next commit.

CHANGE: change must be noted in the release notes.