Added markuptools.make_centered_title_markup().
The function acts as quick way to build up a relatively complex piece of LilyPond markup.

Example:

    abjad> markup = markuptools.make_centered_title_markup('String Quartet', font_size=20)
    abjad> f(markup)
    \markup { \column {
                \center-align {
                    \override #'(font-name . "Times")
                    \fontsize #20 {
                        " "   " "   " "   " "   " "
                        \line { "String Quartet" } 
                        " "   " "   " "
                    }
                }
            } }

The implementation isn't the cleanest because the code is basically just a substitution into
a manually formatted raw string.

But it works for now and can be cleaned up if implemented a tighter way of building markup objects.

Seems like a good pattern to allow into the codebase so long as all the variable parameters
in the markup string are given as input parameters to the function. (Here these are title,
font name and font size.)