Notes on removing the abjad/templates directory.
The results of the previous study ignored the contents of abjad/docs.
This commit repeats the previous study with abjad/docs included.

Results:

1. coventry.ly is not used to produce the docs.

2. lagos.ly is not used to produce the docs.

3. tangiers.ly is not used to produce the docs.

4. oedo.ly is used to produce the following doc pages:
    * docs/chapters/developer_documentation/abjad_book
    * docs/chapters/appendices/history
    * docs/chapters/reference_manual/chords
    * docs/chapters/reference_manual/durations
    * docs/chapters/reference_manual/lilypond_comments
    * docs/chapters/reference_manual/notes
    * docs/chapters/reference_manual/rests
    * docs/chapters/tutorial/parentage
    * docs/chapters/tutorial/threads

5. paris.ly is used to produce the following doc pages:
    * docs/chapters/reference_manual/named_chromatic_pitch

6. thebes.ly is used to produce the following doc pages:
    * docs/chapters/appendics/text_alignment

7. tirnaveni.ly is used to produce the following doc pages:
    * demos/desordre
    * demos/rai
    * docs/chapters/examples/ligeti

Conclusion: the conventry.ly, lagos.ly and tangiers.ly templates can be removed
with no change to the docs. The remaining templates are being used, though unequally.

Strategy: Because oedo.ly and tirnaveni.ly are being used most widely in the docs,
we will try to migrate usage of paris.ly and thebes.ly to either oedo.ly or tirnaveni.ly.
At that point only oedo.ly and tirnaveni.ly will be used in the docs.
Once that is the case, we'll see if we can replace both oedo.ly and tirnaveni.ly with
yet-to-be-created lilypondfiletools. The lilypondfiletools will work like
lilypondfiletools.make_basic_lilypond_file(). Perhaps we will have something like
lilypondfiletools.make_lilypond_file_doc_example(). Perhaps we will also
have an additionally lilypondfiletools.make_lilypond_file_rhythm_example().
At that point we'll remove the abjad/templates diretory entirely.

What to do about users who may by using the abjad/templates directory?
It seems unlikely that this is the case. Precisely the problem with this directory
is that the files it contains are *static*. So users would have to be including the template
files and then making a whole bunch of other LilyPond settings besides.
So probably what we'll do is add ...

    * lilypondfiletools.make_lilypond_file_doc_example()

... and just remove abjad/templates and tell users that the files have been deprecated.