The following two functions no longer default to printing status after completion:
    * iotools.write_expr_to_ly()
    * iotools.write_expr_to_pdf()

The reason can be shown with an example.

The docs contain many examples like this:

    abjad> notes = notetools.make_repeated_notes(1, Fraction(1, 4))
    abjad> staff = stafftools.RhythmicStaff(notes)
    abjad> leaves = staff.leaves
    abjad> markuptools.Markup('XX', 'up')(leaves[0])
    abjad> staff.override.text_script.color = 'red'
    abjad> show(staff)

Previously abjad-book wasn't smart enough to capture output from stdout and stderr.
But now that abjad-book is smart enough to capture such output, we get this in the docs:

    abjad> notes = notetools.make_repeated_notes(1, Fraction(1, 4))
    abjad> staff = stafftools.RhythmicStaff(notes)
    abjad> leaves = staff.leaves
    abjad> markuptools.Markup('XX', 'up')(leaves[0])
    abjad> staff.override.text_script.color = 'red'
    LilyPond file written to 'default-alignment.ly' ...
    abjad> show(staff)

This is a problem because "LilyPond file written to 'default-alignment.ly' ..." is
output from behind-the-scenes rendering of the musical example.

Three solutions are available.

The first is to change every occurence of iotools.write_expr_to_ly() in the docs
to include the keyword setting 'print_status=False'. There are 222 such occurrences.

The second is to temporarily change the default value of iotools.write_expr_to_ly()
print_status keyword when building the docs and then to return the default value
after doc building completes.

The third is to permanently change the value of iotools.write_expr_to_ly() print_status.

The third solution is the one taken here. It's fine for the function to behave in less verbose way.