Added new formattools package.
The package is composer-safe but intended for internal use so that formatting may
remain behind-the-scenes from the perspective of the user.

The package contains the following functions:

    formattools.get_comment_format_contributions
    formattools.get_context_mark_format_contributions
    formattools.get_context_setting_format_contributions
    formattools.get_grob_override_format_contributions
    formattools.get_grob_revert_format_contributions
    formattools.get_lilypond_command_mark_format_contributions

These are, in fact, the core functions involved in formatting every score component.

The functions do not yet have usage examples; those will be added later.