Spanners are now callable.
This means that Spanner now implements __call__.

The method is provided as an experimental way of unifying mark and spanner attachment syntax.

Example:

    abjad> staff = Staff("c'8 d'8 e'8 f'8")

    abjad> beam = spannertools.BeamSpanner()
    abjad> beam(staff[:])
    Staff{4}

    abjad> f(staff)
    \new Staff {
        c'8 [
        d'8
        e'8
        f'8 ]
    }

Note that this works the same way as marks:

    abjad> marktools.Articulation('.')(staff[1])
    Articulation('.')(d'8)

    abjad> f(staff)
    \new Staff {
        c'8 [
        d'8 -\staccato
        e'8
        f'8 ]
    }

This means that there are now two different ways of extending spanners.

The real question will be whether we will want to eventually deprecate
the original (and older) way of extending spanners at initialization.

That is, it may be cleanest to insist that spanners follow the two-step
initialize-then-attach pattern that all marks follow. This two-step pattern
may seem less convenient. But I think it's probably cleaner. So at some point
we'll have to decide whether we want to deprecate the older pattern (and 
thereby break essentiallly all existing spanner examples).