Renames:
    componenttools.list_improper_contents_of_component_that_cross_prolated_offset
    componenttools.list_leftmost_components_with_prolated_duration_at_most

    ==>

    componenttools.get_improper_contents_of_component_that_cross_prolated_offset
    componenttools.get_leftmost_components_with_prolated_duration_at_most

Both 'get' and 'list' are valid verbs with which many different API functions begin.

But there a difference in meaning has evolved in the structure of the API.
Functions named starting with 'get' are in general more complex whereas
functions named starting with 'list' imply a straightforward, flat iteration
of their input arguments.