Added new 'docs' keyword to show().
The keyword is boolean and set to false by default.

Set to true to cause input to be wrapped with documentationtools.make_reference_manual_lilypond_file()
prior to being shown.

Added same 'docs' keyword to iotools.write_expr_to_ly().

These affordances together allow for the replacement of all occurrences of the
deprecated 'oedo.ly' template in the docs.

This commit migrates the reference manual chapter on LilyPond comments
over to the new 'docs' keyword as a test case.