CHANGE:
Previously these four thing ...

    Context.engraver_consists
    Context.engraver_removals
    ContextBlock.engraver_consists
    ContextBlock.engraver_removals

... were (unordered) SETS.

As of this commit, all four things are now (ordered) LISTS.

This changes end-user syntax.

You must now say ...

    staff.engraver_removals.APPEND('Forbid_line_break_engraver')

... rather than saying ...

    staff.engraver_removals.ADD('Forbid_line_break_engraver')

... which will no longer work.

Explanation: it turns out that there are very tricky situations in LilyPond
context definitions where THE ORDER of the appearance of \consists commands
matters for what LilyPond does.

The motivating case to this commit was the polymetric music in the Ligeti Desordre example.

Specifically this context definition ...

   \context {
      \Staff
      \consists Timing_translator
      \consists Default_bar_line_engraver
   }

... is completely different from this context definition ...

   \context {
      \Staff
      \consists Default_bar_line_engraver
      \consists Timing_translator
   }

... in that the first one will draw barlines and the second one won't!

So the order of LilyPond \consists commands is taken as significant by LilyPond.