Extended ContextMark.attach().
The method now checks to make sure that no conflicting context mark is found at *attachment time*.

This prevents the annoying pattern of attaching conflicting tempo marks to components with
the same start time residing in two or three different staves (which then raises an exception
at format-time).

The effective changed was to change this two-line check implemented in ContextMark.attach() ...

        if contexttools.is_component_with_context_mark_attached(start_component, klasses):
            raise ExtraMarkError('component already has context mark attached.')

... to this five-line check ...

        effective_context_mark = contexttools.get_effective_context_mark(start_component, klasses)
        if effective_context_mark is not None:
            if effective_context_mark.start_component.start == start_component.start:
                raise ExtraMarkError(
                    'effective context mark already attached to component starting at same time.')

... instead.

The check is extremely straightforward: we look before we leap to see if an effective context
mark of the appropriate type is already attached to a component elsewhere in the score that
starts at exactly the same time as we would like our new context mark to start.

This works in all cases except time signature marks.

Why? Because calling contexttools.get_effective_context_mark(start_component, TimeSignatureMark) 
results in infinite recursion.

Why? Because time signature marks are treated in some really weird ways in other parts of the
codebase. Without more careful analysis, it looks like the special-cased block
in contexttools.get_effective_context_mark() is probably the problem.

The solution adopted here is to implement the more-strict attachment check on ContextMark
while preserving the older less-strict attachment check for TimeSignatureMark.

If the special-casing for time signature getting is ever completely removed, then the special
implementation of TimeSignatureMark.attach() can also be removed and allow the class to fall
back on the new implementation of ContextMark.attach() committed here.

This closes Google #100.