Added:
    * sequencetools.partition_sequence_by_backgrounded_weights()

The function requires a number of example to understand:

    abjad> sequencetools.partition_sequence_by_backgrounded_weights([-5, -15, -10], [20, 10])
    [[-5, -15], [-10]]

    abjad> sequencetools.partition_sequence_by_backgrounded_weights([-5, -15, -10], [5, 5, 5, 5, 5, 5])
    [[-5], [-15], [], [], [-10], []]

    abjad> sequencetools.partition_sequence_by_backgrounded_weights([-5, -15, -10], [1, 29])
    [[-5], [-15, -10]]

    abjad> sequencetools.partition_sequence_by_backgrounded_weights([-5, -15, -10], [2, 28])
    [[-5], [-15, -10]]

    abjad> sequencetools.partition_sequence_by_backgrounded_weights([-5, -15, -10], [1, 1, 1, 1, 1, 25])
    [[-5], [], [], [], [], [-15, -10]]

Notionally what’s going on here is that the elements of weights are acting as a list of 
successive time intervals into which the elements of sequence are being fit in accordance 
with the start offset of each sequence element.

The function models the grouping together of successive time spans according to 
which of an underlying sequence of time intervals it is in which each time span begins.