Added experimental/handlers package.
The package contains the abstract Handler class.

The package also contains three subpackages:

    handlers.articulations
    handlers.dynamics
    handlers.pitch

The classes implemented in handlers.pitch is truly experimental and hasn't yet
been used in realworld scores. So these classes ...

    handlers.pitch.DiatonicClusterHandler
    handlers.pitch.OctaveTranspositionHandler
    handlers.pitch.PitchHandler
    handlers.pitch.TimewisePitchClassHandler

... can be thought of as stubs checked in for later development.

The classes implemented in handlers.articulations and handlers.dynamics have been used
in realworld scores.

Here are the articulation handlers:

    handlers.articulations.ArticulationHandler
    handlers.articulations.PatternedArticulationsHandler
    handlers.articulations.ReiteratedArticulationHandler
    handlers.articulations.RepeatedMarkupHandler
    handlers.articulations.StemTremoloHandler

And here are the dynamic handlers:

    handlers.dynamics.DynamicHandler
    handlers.dynamics.NoteAndChordHairpinHandler
    handlers.dynamics.NoteAndChordHairpinsHandler
    handlers.dynamics.ReiteratedDynamicHandler
    handlers.dynamics.TerracedDynamicsHandler

So what is a handler?

The name is rather thin semantically: all that is denoted is a agentive thing. That is,
a thing that does something.

(This is, after all, what is denoted by English nouns ending in -er, -or, -ar: 
butcher, baker, time-token maker.)

So the handlers simply represent the move away from vanilla API *functions* and
towards more sophisticated *classes* that implement __call__.

In this respect handlers shared much with time-token makers. And, in fact, the time-token
makers are basically RhythmHandlers. (Although no such class actually exists.)

The handlers give you object oriented access to a function calls that transform some bit of score.
And, depending on implementation, handlers might also give you statal access to some
sequence of numbers, pitches, articulations, dynamics or the like. So handlers also share
a great deal in common with the idea of the statal server (current implemented in specificationtools).
One of the things that will need to be worked out is the future relationship between handlers
and statal servers of all types.

There are some handler test files. These test files show handlers in action.

Here is one handler test:

    pattern = handlers.articulations.PatternedArticulationsHandler([['>', '-'], ['.']])
    staff = Staff("c'8 d'8 r8 e'8 f'8 r8 g'8 r8")
    pattern.apply(staff)

    f(staff)
    \new Staff {
        c'8 -\accent -\tenuto
        d'8 -\staccato
        r8
        e'8 -\accent -\tenuto
        f'8 -\staccato
        r8
        g'8 -\accent -\tenuto
        r8
    }

The handler applies a repeating pattern to all notes-and-or-chords in an expression, skipping over rests.

Notice here that handler implements a (somewhat unfortunately named) apply() method rather than __call__().
I don't remember exactly what the reason for this is at the moment; I think it has something to
do with DynamicHandler objects implementing __call__ to allow the resetting of internal values.
(So this will be a design consideration that will have to be dealt with, too: what will be
the long-term meaning of __call__ implemented on these objects? There should only be one such pattern.)

Here's another example from the handler test files:

    hairpin = handlers.dynamics.NoteAndChordHairpinHandler(('p', '<', 'f'))
    staff = Staff("r4 c'8 d'8 r4 e'8 r8")
    hairpin.apply(staff)

    \new Staff {
        r4
        c'8 \< \p
        d'8
        r4
        e'8 \f
        r8
    }

The handler finds groups of consecutive note-and-or-chords and applies a piano-to-forte crescendo 
to each group.