Added SegmentSpecification.set_retrograde_divisions().
These two methods provide a similar interface and work in similar ways:

    SegmentSpecification.set_rotated_divisions()
    SegmentSpecification.set_retrograde_divisions()

Both methods basically say "go fetch divisions from somewhere in the score;
then, after you've fetched divisions, postprocess them in some interesting way."

In the case of the first method, the postprocessing is rotation.

In the case of the second method, the postprocessing retrogression.

The implementation of both functions is concise. Both functions rely on a DivisionRetrievalRequest
to store values specified by the composer.

Many more such named methods are of course possible. So an open design question will be how
many such fetch-then-postprocess named methods we'll want in the public API. It is certainly
possible to pollute the public API with too many such named methods. But a core number of
such named methods seems like a very good idea. So it seems like there will be something
of a Goldilocks situation here.

Also note that both methods are merely wrappers around the more general SegmentSpecification.set_divisions().
This more general method provide a 'callback' keyword that already allows for any 
type of postprocessing whatever. So SegmentSpecification.set_rotated_divisions() and
SegmentSpecification.set_retrograde_divisions() are, in a sense, completely unnecessary.
But there is no substitute for having the conceptually tight interfaces provided by the
more specific pair of methods.

In fact, it seems likely that much of the work of implementing specificationtools will be
the production of interface sugar. For the very important reason of streamlining the compositional
act of specification as much as possible.

At any rate, this pair of functions constitutes something of a pattern that we can return to
as we implement further.

An example of using the generalized 'callback' keyword will be provided in a future commit,
for completeness' sake and also to make sure the affordance is tested in isolation.