Added custom PartIndicator class to specificationtools.
The class brings together a ratio with an indication of specially chosen parts of the ratio.

Example:

    >>> specificationtools.PartIndicator((1, 2, 1), 0)
    PartIndicator(Ratio(1, 2, 1), 0)

This models is a 1:2:1 partition (of some as-of-yet unspecified thing) together
with the indication to take the part at index from the resulting partition.

Compare with:

    >>> specificationtools.PartIndicator((1, 2, 1), slice(1, 3))
    PartIndicator(Ratio(1, 2, 1), slice(1, 3, None))

This models a 1:2:1 partition together with the indication to take the last two
parts from the resulting partition.