Wrote abjadbooktools.AbjadBook class.
This class wraps almost all functionality of the old abjad-book script.

Each part of the process is broken down into a different private method, bound to the class.

As these private methods do not internally reference any of the properties of the class, they can be tested atomically (tests are forthcoming).

These methods are, however, not useful outside of the scope of the abjad-book process, so it seems unwise to export them as public functions.

A new version of the abjad-book script, to be staged in experimental/scr is also forthcoming.