Added built-in 'left', 'right', 'up' and 'down' constants.
These load into __builtins__ on specificationtools import.

Here are the specificationtools/__init__.py lines that do this:

    __builtins__['left'] = VectorConstant('x', -1, 'left')
    __builtins__['right'] = VectorConstant('x', 1, 'right')
    __builtins__['up'] = VectorConstant('y', 1, 'up')
    __builtins__['down'] = VectorConstant('y', -1, 'down')

The specificationtools.VectorConstant class is new.

The class models a dimensioned constant with user-specifiable repr.

Example:

        >>> left = specificationtools.VectorConstant('x', -1, 'left')

        >>> left
        left

Defining constants like this and then loading into __builtins__ gives us the ability
to make constants on par with True, False and None.

The 'left' and 'right' constants compare with each other.
The 'up' and 'down' constants compare with each other.
But 'left' cannot compare with 'up'. This is accomplished by initializing the dimension
of the constants differently.

                                                   * * *

What motivated the introduction of vector constants?

The ability to say edge=left and edge=right when initializing TemporalCursor objects.

Note that you now say things like ...

        >>> specificationtools.TemporalCursor(edge=right)
        TemporalCursor(edge=right)

... to set a cursor at the right edge of the entire score.

Which seems very clean.

Yes, it would have been possible to use strings to say edge='right' and edge='left'.

But strings supplied as values to keyword parameters is avoided this way.
And it seems that the concepts of left and right are comparable in essentially every way to true and false.