documentationtools.ModuleCrawler can now automatically determine the root package name of the directory it is asked to begin crawling in.
This means that the experimental directory can now be crawled without error (by various developer scripts, for example).

As a side-effect, documentationtools.ModuleCrawler can only begin crawling in directories containing which are also Python package directories (that is, which contain __init__.py files).

This limitation may be lifted in the future, if necessary.  For now, it functions appropriately in all common use cases.

documentationtools.APICrawler now also prints the names of the files it updates, instead of merely the files it creates.