Moved all tools packages marked "additional composition packages" in the API to the main section of the API (labeled "Abjad composition packages").
The idea of the "additional composition packages" moniker is now dead.

The Abjad API now supports only a two-way distinction:

    Abjad composition packages / internals packages

The idea of the "unstable" package is deprecated now, too.

Unstable work should be developed in the experimental branch.

The tonalitytools package remains as the sole package listed as unstable.

The tonalitytools package will be moved to the experimental branch in an upcoming commit.