Fixed lexical state conflict bug in LilyPondParser.
The solution was relatively painless, and hinged on two small changes.

First, a new private method was added to the LilyPondParser class itself::

    def _relex_lookahead(self):
        difference = self._parser.lookahead.lexpos - self._lexer.lexpos
        self._lexer.skip(difference)
        self._parser.lookahead = None 

This method calculates the number of characters between the lexer's current position and the position of the token currently held as lookahead (meaning, the token waiting to be parsed).

It then tells the lexer to skip _backwards_ to the position where the lookahead token starts.

Finally, it overwrites the lookahead token with None, which will force the parser to ask for a new token the next time it needs one, effectively relexing the old lookahead token.

Second, inside the _LilyPondSyntacticalDefinition class (where all syntax rules reside), we make one tiny change in a markup rule::

    def p_full_markup__MARKUP__markup_top(self, p):
        'full_markup : MARKUP markup_top'
        p[0] = markuptools.Markup(p[2])
        self.client._lexer.pop_state()
        self.client._relex_lookahead()

Here, we simply ask for relexing immediately after popping the lexer's current lexical state.

And that's how it's fixed.