Updated 2.9 optimization log with additional tests for marks and spanner.
On a staff of 100 notes, the general pattern seems to be about 10,000 additional
functions calls and 0.01 sec additional compute time for 100 new marks.

The pattern also seems to be about 10,000 additional function calls and 0.01 sec
additional compute time for a beam spanner that encompasses 100 notes.

These results were all run on 2.9 r6023.