Wrote more observations about timespan-relative (or timespan-reference) selection.
There appear to be 13 ways to pick out sets of positively durated score objects with
respect to an arbitrary positively durated reference timespan.

The document committed here that enumerates these constitutes something of a proposal
for a formal model of the "fringe" that can result when working with timespans
that are allowed to overlap.