Final timespan relations.
One compound inequality was missing from the previous revision of the catalog.

There are 27 relations possible between timespans AB and XY:

  [1] A <  X                        "measures that start before blah starts"
  [2] A == X                        "measures that start when blah starts"
  [3] X <  A                        "measures that start after blah starts"

  [4] A <  Y                        "measures that start before blah stops"
  [5] A == Y                        "measures that start when blah stops"
  [6] Y <  A                        "measures that start after blah stops"

  [7] B <  X                        "measures that stop before blah starts"
  [8] B == X                        "measures that stop when blah starts"
  [9] X <  B                        "measures that stop after blah starts"

 [10] B <  Y                        "measures that stop before blah stops"
 [11] B == Y                        "measures that stop when blah stops"
 [12] Y <  B                        "measures that stop after blah stops"

 [13] A <  X && X <  B              "measures that start before blah starts and stop after blah starts"

 [14] A <  X && B <  Y              "measures that start before blah starts and stop before blah stops"
 [15] A <  X && B == Y              "measures that start before blah starts and stop when blah stops"
 [16] A <  X && Y <  B              "measures that start before blah starts and stop after blah stops"

 [17] A == X && B <  Y              "measures that start when blah starts   and stop before blah stops"
 [18] A == X && B == Y              "measures that start when blah starts   and stop when blah stops"
 [19] A == X && Y <  B              "measures that start when blah starts   and stop after blah stops"

 [20] X <  A && A <  Y              "measures that start after blah starts  and start before blah stops"

 [21] X <  A && B <  Y              "measures that start after blah starts  and stop before blah stops"
 [22] X <  A && B == Y              "measures that start after blah starts  and stop when blah stops"
 [23] X <  A && Y <  B              "measures that start after blah starts  and stop and after blah stops"

 [24] A <  Y && Y <  B              "measures that start before blah stops  and stop after blah stops"

 [25] X <  B && B <  Y              "measures that stop after blah starts   and stop before blah stops"

 [26] X <  A && A <  Y && Y <  B    "measures that start after blah starts  and start before blah stops
                                                                            and stop after blah stop"
 [27] A <  X && X <  B && B <  Y    "measures that start before blah start  and stop after blah starts
                                                                            and stop before blah stops"

The problem was worked out essentially with brute force. How? Observe that the first 12 inequalities
are all simple: each contains one only comparison operation. All other inequalities must be
compounded from this inventory of 12 simple inequalities. Working the problem then essentially
reduced to checking the permissibility of each of the 12 simple equalities with each of the 11
other simple inequalities. One way to visualize this is as K12 the complete graph on 12 vertices.
This graph contains (12 - 11) / 2 == 66 edges. So building this enumeration was essentially the
same as coloring each of the 66 edges in K12 either blue (for an acceptable combination of
simple inequalities) or red (for an unacceptable combination of simple inequalities). Of the 66
edges, found 13 to be acceptable (and the remaining 53 to be unacceptable). This brings us
so far to 12 simple inequalities and 13 inequalities comprising exactly two parts. This gives
us 12 + 13 = 25 inequalities modeling 25 timespan relations. But the catalog is complete only
when we add 2 inequalities comprising exactly *three* parts.

Visualized this way, a powerful generalization is available: the 27 relations obtaining
between arbitrary timespans AB and XY are isomorphic to exactly *THE CLIQUES OF K12*
colored for acceptability as described above.